feat: add spec view
