+++ packages/mosaic/src/charts/MosaicChartSettingsPanel.tsx
@@ -1,8 +1,6 @@
-import {type FC, useCallback, useState} from 'react';
+import {type FC} from 'react';
 import {MosaicChartSettings} from './chart-settings/MosaicChartSettings';
-import {MosaicChartSpecViewerPanel} from './chart-settings/MosaicChartSpecViewerPanel';
 import {type ChartConfig} from './chart-types/chart-config';
-import {useMosaicChartRenderContext} from './useMosaicChartRenderContext';
 import {DataTable} from '@sqlrooms/db';
 
 export type MosaicChartSettingsPanelProps = {
@@ -16,23 +14,6 @@ export const MosaicChartSettingsPanel: FC<MosaicChartSettingsPanelProps> = ({
   config,
   onChange,
 }) => {
-  const [viewMode, setViewMode] = useState<'settings' | 'spec'>('settings');
-
-  const renderContext = useMosaicChartRenderContext(dataTable, config);
-
-  const handleBackToSettings = useCallback(() => {
-    setViewMode('settings');
-  }, []);
-
-  if (renderContext.type === 'spec' && viewMode === 'spec') {
-    return (
-      <MosaicChartSpecViewerPanel
-        spec={renderContext.spec}
-        onBack={handleBackToSettings}
-      />
-    );
-  }
-
   const columns = dataTable?.columns || [];
 
   return (
